added limit for room capacity

+++ server/routes/rooms.js
@@ -7,12 +7,18 @@ require("dotenv").config();
 const router = express.Router();
 
 
+MAX_MEMBERS = 8
 async function handleRoom(req, res, update) {
-    const updatedRoom = await Room.findOneAndUpdate(
-        { code: req.body.code },
+    const room = await Room.findOne({ code: req.body.code });
+    if (!room) return res.status(400).json({ error: "A room with that code does not exist!" });
+    if (room.members.length >= MAX_MEMBERS) return res.status(400).json({ error: "Sorry, that room is full!" });
+
+    const updatedRoom = await Room.findByIdAndUpdate(
+        { _id: room._id },
         update,
         { new: true }
     );
+
     if (!updatedRoom) return res.status(400).json({ error: "A room with that code does not exist!" });
     if (updatedRoom.members.length === 0) await updatedRoom.deleteOne();
     res.status(200).json(updatedRoom);
